@@ -1,44 +1,45 @@
 import React from 'react';
-import Producto from './Componentes/Producto';
+import './App.css';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import Producto from './Componentes/Producto'
 import { Container, Row } from 'reactstrap';
 import Navegacion from './Componentes/Navegacion';
-import './App.css';
-import lista from './listaProductos.json';
+import listaProductos from "./listaProductos.json";
+console.log();
 
 class App extends React.Component {
-  constructor(){
-    const {listaProductos} = lista
+  constructor() {
     super();
-
     this.state = {
       listaProductos
-    };
+    }
   }
 
   render() {
-    const arregloComponente = this.state.listaProductos.map(
+    const arregloComponentes = this.state.listaProductos.map(
       (listaProductos, i) => {
-        return (
+        return(
           <Producto 
-            key = {i}
+            key={i}
             titulo = {listaProductos.titulo}
             imagen = {listaProductos.imagen}
             descripcion = {listaProductos.descripcion}
             precio = {listaProductos.precio.replace(".", "")}
             stock = {listaProductos.stock}
-            />
-            )
-          });
+          />
+        );
+      }
+    );
 
     return (
       <Container>
-        <Navegacion titulo="Mi primer sitio de compras en React" />
+        <Navegacion titulo="Mi primer sitio de compras en React"/>
         <Row>
-          {arregloComponente}
+          {arregloComponentes}
         </Row>
       </Container>
     );
-  } 
+  }
 }
 
 export default App;
