@@ -1,15 +1,15 @@
 import React from "react";
-import { Badge, Button, Popover, PopoverBody, PopoverHeader, Table } from "reactstrap"; 
-import {listaCarrito} from '../listaCarrito.json';
+import { Table, PopoverBody, PopoverHeader, Popover, Button, Badge } from 'reactstrap';
+import listaCarrito from '../listaCarrito.json';
 
 class Carro extends React.Component {
     constructor() {
         super();
         this.state = {
-            popoverOpen:false,
+            popoverOpen: false,
             listaCarrito
-        };
-        
+        }
+
         this.toggle = this.toggle.bind(this);
     }
 
@@ -17,63 +17,70 @@ class Carro extends React.Component {
         this.setState(prevState => ({
             popoverOpen: !prevState.popoverOpen
         }));
-
     }
 
     sumaTotal() {
         let total = 0;
-        let sumaTotal = this.state.listaCarrito.map(
+        this.state.listaCarrito.map(
             (listaCarrito) => {
-                return total+= parseInt(listaCarrito.precio);
+                return total += parseInt(listaCarrito.precio);
             }
         )
-        return(total)
+        return(total);
     }
-     
+
     render() {
         const arregloCarrito = this.state.listaCarrito.map(
             (listaCarrito, i) => {
                 return(
-                    <tr>
+                    <tr key={i}>
                         <td>{(i += 1)}</td>
                         <td>{listaCarrito.titulo}</td>
                         <td>{listaCarrito.precio}</td>
                     </tr>
                 );
             }
-        )
-        return (
+        );
+        return(
             <div>
-                <Button id="Popover1" color="info">
+                <Button id="Popover1" color="primary">
                     <span className="material-icons">shopping_cart</span>
-                    <Badge color="secondary" id="Badge1">{listaCarrito.length}</Badge>
+                    <Badge color="primary" id="Badge1">{listaCarrito.length}</Badge>
                 </Button>
-                <Popover target="Popover1" placement="bottom" isOpen={this.state.popoverOpen} toggle={this.toggle}>
+                <Popover 
+                    target="Popover1" 
+                    placement="bottom" 
+                    isOpen={this.state.popoverOpen} 
+                    toggle={this.toggle}
+                >
                     <PopoverHeader>Listado de compras</PopoverHeader>
                     <PopoverBody>
-                        <Table className="table table-bordered table-sm">
-                            <thead>
+                        <Table>
+                           <thead>
                             <tr>
                                 <th>#</th>
                                 <th>Producto</th>
-                                <th>Precio Producto</th>
+                                <th>Precio</th>
                             </tr>
-                            </thead>
-                            <tbody>
+                           </thead>
+                           <tbody>
                                 {arregloCarrito}
-                            </tbody>
-                            <tfoot>
+                           </tbody>
+                           <tfoot>
                                 <tr>
-                                    <td></td>
+                                    <td>{arregloCarrito.length}</td>
                                     <td>Total:</td>
-                                    <td>$ {Intl.NumberFormat("de-DE").format(this.sumaTotal())} CLP</td>
+                                    <td>$ <b>{Intl.NumberFormat("de-DE").format(this.sumaTotal())} CLP</b></td>
                                 </tr>
-                            </tfoot>
+                           </tfoot>
                         </Table>
                     </PopoverBody>
                 </Popover>
+                
             </div>
         );
     }
+
 }
+
 export default Carro;
