@@ -1,25 +1,32 @@
 import React from "react";
-import {CardSubtitle, CardTitle, CardBody,CardImg, Col, Card, CardText} from 'reactstrap';
-import './Producto.css'
-import FichaProducto from "./FichaProducto";
+import { 
+    Card, 
+    CardBody, 
+    CardImg, 
+    CardSubtitle, 
+    CardText, 
+    CardTitle, 
+    Col 
+} from "reactstrap";
+
+import FichaProducto from './FichaProducto';
+import './Producto.css';
 
 class Producto extends React.Component {
     render() {
         return (
             <Col sm="4">
-                <Card className="Card">
-                    <CardImg src={this.props.imagen} />
+                <Card className="Card" body outline color="primary">
+                    <CardImg src={this.props.imagen} alt={this.props.titulo} />
                     <CardBody>
-                        <CardTitle><b>{this.props.titulo}</b></CardTitle>
-                        <CardSubtitle><b>Valor: </b> {this.props.precio}</CardSubtitle>
-                        <CardText>
-                            {this.props.descripcion}
-                        </CardText>
-                        <FichaProducto props={this.props} />
+                        <CardTitle>{this.props.titulo}</CardTitle>
+                        <CardSubtitle>Valor:$ <b>{Intl.NumberFormat("de-DE").format(this.props.precio)}</b> CLP</CardSubtitle>
+                        <CardText>{this.props.descripcion}</CardText>
+                        <FichaProducto props={this.props}/>
                     </CardBody>
                 </Card>
             </Col>
-        )
+        );
     }
 }
 
